add some tests around creating new contacts and fix create check
For some reason, at least in the test setup, the error that
Salesforce returns on upsert with the Last Name field missing lists
the API name, not the field label as previously observed. So this
may or may not fix any real bug, but it does make the added test
pass.