fix updating sync status when subscriber e-mail changes
With this change, we add a second Contact ID field to MailchimpSyncStatus__c,
this one as an external lookup field. That way, we can upsert the sync
statuses on the contact ID (which we now should have in all cases) instead of
the e-mail address. This should ensure that a contact only ever has one
associated MailchimpSyncStatus__c record.